requirements.txt
Flask
requests
tiktoken
PyYAML
pytest
gunicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
Flask-Mail
openai
pydub
bleach
Flask-SQLAlchemy
psycopg2-binary
werkzeug
itsdangerous
google-cloud-storage
cachetools
python-dotenv